Create FinalProjectXX
This code worked in the first few runs, but some part of it breaks after the initial run.  Around the part with the string list, it gets an exception error for the word list and claims it cant read it.  Debugging or building caused the file to be un-openable sometimes.